+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Added` for new features
 
+- argument `--sheets` allows for Comma-separated list of sheet names to compare (default: all shared and unique sheets)
+
 ### `Changed` for changes in existing functionality
 
 - GitHub Action polish-the-code.yml chains prettier-fix and only then super-linter
@@ -20,8 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### `Fixed` for any bugfixes
 
 - minor pylint identified issues
-  - xlsx_compare.py:24:5: W0511: Todo potential to speed up the comparison (fixme)
-  - xlsx_compare.py:81:9: W0511: Todo Add a function to save the differences to an Excel file (fixme)
+  - xlsx_compare.py:24:5: W0511: todo potential to speed up the comparison (fixme)
+  - xlsx_compare.py:81:9: W0511: todo Add a function to save the differences to an Excel file (fixme)
   - xlsx_compare.py:1:0: C0114: Missing module docstring (missing-module-docstring)
   - xlsx_compare.py:12:0: C0115: Missing class docstring (missing-class-docstring)
   - xlsx_compare.py:12:0: R0903: Too few public methods (0/2) (too-few-public-methods)